api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
sqlmodel==0.0.8
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

